+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_update_backup_policy_request.py
@@ -15,6 +15,7 @@ class UpdateBackupPolicyRequest(DaraModel):
         cold_data_backup_interval: int = None,
         cold_data_backup_retention: int = None,
         cross_region_data_backup_retention: int = None,
+        cross_region_filter_value: str = None,
         cross_region_log_backup_retention: int = None,
         dbinstance_name: str = None,
         dest_cross_region: str = None,
@@ -36,6 +37,7 @@ class UpdateBackupPolicyRequest(DaraModel):
         self.cold_data_backup_interval = cold_data_backup_interval
         self.cold_data_backup_retention = cold_data_backup_retention
         self.cross_region_data_backup_retention = cross_region_data_backup_retention
+        self.cross_region_filter_value = cross_region_filter_value
         self.cross_region_log_backup_retention = cross_region_log_backup_retention
         # This parameter is required.
         self.dbinstance_name = dbinstance_name
@@ -82,6 +84,9 @@ class UpdateBackupPolicyRequest(DaraModel):
 
         if self.cross_region_data_backup_retention is not None:
             result['CrossRegionDataBackupRetention'] = self.cross_region_data_backup_retention
+
+        if self.cross_region_filter_value is not None:
+            result['CrossRegionFilterValue'] = self.cross_region_filter_value
 
         if self.cross_region_log_backup_retention is not None:
             result['CrossRegionLogBackupRetention'] = self.cross_region_log_backup_retention
@@ -146,6 +151,9 @@ class UpdateBackupPolicyRequest(DaraModel):
 
         if m.get('CrossRegionDataBackupRetention') is not None:
             self.cross_region_data_backup_retention = m.get('CrossRegionDataBackupRetention')
+
+        if m.get('CrossRegionFilterValue') is not None:
+            self.cross_region_filter_value = m.get('CrossRegionFilterValue')
 
         if m.get('CrossRegionLogBackupRetention') is not None:
             self.cross_region_log_backup_retention = m.get('CrossRegionLogBackupRetention')